+++ db/migrate/20260908000001_add_serial_number_to_things.rb
@@ -1,0 +1,6 @@
+class AddSerialNumberToThings < ActiveRecord::Migration[8.1]
+  def change
+    add_column :things, :serial_number, :string
+    add_index :things, :serial_number, where: "serial_number IS NOT NULL"
+  end
+end